Separate unit and integration tests
