+++ harmony-api/jest.config.js
@@ -3,5 +3,5 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   modulePathIgnorePatterns: ['build', 'node_modules', 'src'],
-  setupFilesAfterEnv: ['<rootDir>/__test__/prisma/singleton.ts'],
+  setupFilesAfterEnv: ['<rootDir>/__test__/testUtils/prisma/singleton.ts'],
 }
